pyproject.toml
[tool.poetry]
name = "fastapu-google-auth"
version = "0.1.0"
description = ""
authors = ["Харицкий Данил Константинович <Khlive.ftf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi = {extras = ["all"], version = "^0.104.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
alembic = "^1.12.1"
asyncpg = "^0.29.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
authlib = "^1.2.1"
httpx = "^0.25.1"
pydantic-settings = "^2.0.3"
random-username = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
